refactor: code refactor for attributes exatractors.
